requirements.txt
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.13
requests==2.31.0
numpy==1.26.4
openai-whisper==20231117
torch==2.2.1
ffmpeg-python==0.2.0
opencc-python-reimplemented==0.1.7
python-dotenv==1.0.0
soundfile==0.12.1
funasr==1.2.6
modelscope==1.13.0
sounddevice==0.4.6
webrtcvad==2.0.10
numpy==1.26.4
fastapi==0.115.12
uvicorn==0.34.0
cn2an==0.5.23